requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
autopep8==1.4.4
backcall==0.1.0
Brlapi==0.6.6
cachetools==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
gast==0.2.2
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpg==1.10.0
grpcio==1.28.1
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imageio==2.8.0
imutils==0.5.3
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.0.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
Markups==2.0.1
MarkupSafe==1.0
matplotlib==3.2.1
meld==3.18.0
netifaces==0.10.4
networkx==2.4
numpy==1.18.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.2.0.34
opt-einsum==3.2.1
parso==0.7.0
pdftotext==2.1.3
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
progressbar2==3.51.1
prompt-toolkit==3.0.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyenchant==2.0.0
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-magic==0.4.16
python-utils==2.4.0
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
system-service==0.3
systemd-python==234
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
textile==3.0.0
torch==1.5.0
tornado==6.0.4
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.3.2
